refactor: move view state and columns logic to useInvoice hook
