+++ frontend/emiratads/app/client/invoice/useInvoice.tsx
@@ -1,5 +1,6 @@
-import { useState, useEffect, useMemo } from "react";
-import { transactionTypeEnum, Transaction, Invoice } from "@/app/types/InvoiceTypes";
+import { useState } from "react";
+import { maskCurrency } from "@/app/utils/currencyMask";
+import { transactionTypeEnum, Invoice } from "@/app/types/InvoiceTypes";
 
 const useInvoice = () => {
     const invoice: Invoice = {
@@ -31,8 +32,31 @@ const useInvoice = () => {
         ]
     };
 
+    const [isCardView, setIsCardView] = useState(true);
+    const toggleView = () => setIsCardView(!isCardView);
+
+    const columns = [
+        {
+            accessorKey: "data",
+            header: "Data",
+            cell: ({ row }: { row: { original: { data: string } } }) => new Date(row.original.data).toLocaleString(),
+        },
+        { accessorKey: "quantidade_milhas", header: "Milhas" },
+        {
+            accessorKey: "valor",
+            header: "Valor em reais",
+            cell: ({ row }: { row: { original: { valor: number } } }) => maskCurrency(row.original.valor),
+        },
+        { accessorKey: "descricao", header: "Descrição" },
+        { accessorKey: "codigo_reserva", header: "Reserva", cell: ({ row }: { row: { original: { codigo_reserva: string } } }) => row.original.codigo_reserva || "-" },
+        { accessorKey: "tipo", header: "Tipo" },
+    ];
+
     return {
-        invoice
+        invoice,
+        isCardView,
+        toggleView,
+        columns
     };
 };
